+++ src/main/java/io/kestra/plugin/hightouch/package-info.java
@@ -3,6 +3,6 @@
     "Hightouch is a reverse ETL product which allows you to easily activate your data.",
     categories = PluginSubGroup.PluginCategory.TOOL
 )
-package io.kestra.plugin.hightouch.connections;
+package io.kestra.plugin.hightouch;
 
 import io.kestra.core.models.annotations.PluginSubGroup;
