+++ src/test/java/io/kestra/plugin/hightouch/SyncInvalidSyncIdTest.java
@@ -1,24 +1,25 @@
-package io.kestra.plugin.hightouch.connections;
+package io.kestra.plugin.hightouch;
 
 import com.google.common.collect.ImmutableMap;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.runners.RunContextFactory;
 import io.micronaut.context.annotation.Value;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
+import io.micronaut.http.client.exceptions.HttpClientResponseException;
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.*;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @MicronautTest
-class SyncTest {
+class SyncInvalidSyncIdTest {
     @Inject
     private RunContextFactory runContextFactory;
 
-    @Value("${hightouch.sync-id}")
-    private Long syncId;
+    @Value("00000000000000000000000")
+    private Long invalidSyncId;
 
     @Value("${hightouch.token}")
     private String token;
@@ -29,13 +30,14 @@ class SyncTest {
 
         Sync task = Sync.builder()
                 .token(this.token)
-                .syncId(this.syncId)
+                .syncId(this.invalidSyncId)
                 .build();
 
-        Sync.Output runOutput = task.run(runContext);
-
-        assertThat(runOutput, is(notNullValue()));
-        assertThat(runOutput.getRunId(), is(notNullValue()));
-
+        Throwable exception = assertThrows(
+                HttpClientResponseException.class,
+                        () -> task.run(runContext),
+                        "Expected Sync() to throw 404 error with HttpClientResponseException, but it didn't"
+        );
+        assertThat(exception.getMessage(), containsString("Request failed with status '404'"));
     }
 }
